[#2524] extract interface from PinpointCilent
